Backend3/requirements.txt
marshmallow>=3
flask
flask-restful
flask-migrate
flask-sqlalchemy
flask-marshmallow
flask-jwt-extended
marshmallow-sqlalchemy
python-dotenv
passlib
apispec[yaml]
apispec-webframeworks
tox
uwsgi==2.0.20
psycopg2-binary
flask-cors
flask-seeder
minio
pydantic
Pillow
Unidecode
redis
pandas
openpyxl
zipp==3.7.0
pytz
sentry-sdk[flask]
xlrd